+++ twitchapi/chatbot.py
@@ -12,7 +12,7 @@ class ChatBot:
 
     def __init__(self, client_id: str, client_secret: str, bot_name: str, channel_name: str, subscriptions: list[str],
                  redirect_uri_auth: str = REDIRECT_URI_AUTH, timeout=DEFAULT_TIMEOUT, right: list[str] = None,
-                 channel_point_subscription: list[str] = None):
+                 channel_point_subscription: list[str] = None, store_in_db:bool=False):
         self._client_id = client_id
         self.__client_secret = client_secret
 
@@ -57,7 +57,7 @@ class ChatBot:
         self.__event_sub = EventSub(bot_id=self._bot_id, channel_id=self._channel_id,
                                     subscription_types=self.__subscription, auth_server=self.__auth,
                                     trigger_map=self.__trigger_map,
-                                    channel_point_subscription=channel_point_subscription)
+                                    channel_point_subscription=channel_point_subscription, store_in_db=store_in_db)
 
         self.__thread = ThreadWithExc(target=self.__run_event_server)
         self.__thread.start()
@@ -150,7 +150,7 @@ class ChatBot:
     def prediction_end(self, title: str, result: dict, winning_pred: str):
         pass
 
-    def new_ban(self, user_name: str, moderator_name:str, reason: str, start_ban: str, end_date: str, permanent: bool):
+    def new_ban(self, user_name: str, moderator_name:str, reason: str, start_ban: str, end_ban: str, permanent: bool):
         pass
 
     def new_vip(self, user_name: str):